requirements.txt
requests
python-dotenv
pyTelegramBotAPI
google-generativeai
gspread
google-auth
apify-client
